change l'authentification pour le endpoint `refresh`
