+++ import_integration_test.go
@@ -11,10 +11,10 @@ func TestMepHandler(t *testing.T) {
 	//ass := assert.New(t)
 	//rPath := "/utils/mep"
 	//test.Viperize(map[string]string{
-	//	"sourceEntreprise":    "test/emptyJSON.gz",
-	//	"sourceEtablissement": "test/emptyJSON.gz",
+	//	"sourceEntreprise":    "/Users/raphaelsquelbut/Downloads/entreprise_sample.json.gz",
+	//	"sourceEtablissement": "/Users/raphaelsquelbut/Downloads/etablissement_sample.json.gz",
 	//})
-
+	//
 	//response := test.HTTPGet(t, rPath)
 	//ass.Equalf(http.StatusOK, response.StatusCode, "body de la réponse : %s", test.GetBodyQuietly(response))
 }
