+++ src/ops/http.go
@@ -4,16 +4,14 @@ package ops
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
+	"github.com/signaux-faibles/datapi/src/core"
 	"github.com/signaux-faibles/datapi/src/utils"
-	"github.com/spf13/viper"
-	"io"
-	"log"
 	"net/http"
 )
 
 // ConfigureEndpoint configure l'endpoint du package `ops`
 func ConfigureEndpoint(path string, api *gin.Engine) {
-	endpoint := api.Group(path, adminAuthMiddleware())
+	endpoint := api.Group(path, core.AdminAuthMiddleware)
 	endpoint.GET("/import", importHandler) // 1
 	endpoint.GET("/keycloak", keycloakUsersHandler)
 	endpoint.GET("/metrics", gin.WrapH(promhttp.Handler()))
@@ -23,32 +21,13 @@ func ConfigureEndpoint(path string, api *gin.Engine) {
 	endpoint.GET("/mep/:algo", importHandler, importSireneHandler, importListesHandler)
 }
 
-func adminAuthMiddleware() gin.HandlerFunc {
-	var whitelist = viper.GetStringSlice("adminWhitelist")
-	var wlmap = make(map[string]bool)
-	for _, ip := range whitelist {
-		wlmap[ip] = true
-	}
-
-	return func(c *gin.Context) {
-		if !wlmap[c.ClientIP()] {
-			log.Printf("Connection from %s is not granted in adminWhitelist, see config.toml\n", c.ClientIP())
-			c.AbortWithStatus(http.StatusForbidden)
-			return
-		}
-	}
-}
-
 func importSireneHandler(c *gin.Context) {
 	err := importSirene()
 	if err != nil {
 		utils.AbortWithError(c, err)
 		return
 	}
-	c.Stream(func(w io.Writer) bool {
-		w.Write([]byte("sirenes mis à jour"))
-		return true
-	})
+	c.Next()
 	//c.JSON(http.StatusOK, "sirenes mis à jour")
 }
