+++ src/refresh/http.go
@@ -15,7 +15,7 @@ import (
 
 // ConfigureEndpoint configure le endpoint du package `refresh`
 func ConfigureEndpoint(path string, api *gin.Engine) {
-	refreshRoute := api.Group(path, core.AuthMiddleware(), core.LogMiddleware)
+	refreshRoute := api.Group(path, core.AdminAuthMiddleware)
 	refreshRoute.GET("/start", startHandler)
 	refreshRoute.GET("/status/:uuid", statusHandler)
 	refreshRoute.GET("/last", lastHandler)
